+++ src/components/pages/AppSubscription/AppSubscriptionDetailOverlay/index.tsx
@@ -119,6 +119,10 @@ const AppSubscriptionDetailOverlay = ({
         data?.tenantUrl ?? '',
       ],
       [
+        `${t('content.appSubscription.detailOverlay.appId')}`,
+        data?.appInstanceId ?? 'N/A',
+      ],
+      [
         `${t('content.appSubscription.detailOverlay.technicalName')}`,
         data?.technicalUserData && data.technicalUserData.length > 0
           ? data.technicalUserData[0].name